+++ sources/components/rb__childfx/rb__childfx.js
@@ -110,7 +110,7 @@
 
                     var options = {
                         start: {},
-                        end: Object.assign({}, rb.parsePseudo(elem), that.parseHTMLOptions(elem)),
+                        end: Object.assign({}, rb.parsePseudo(elem), rb.parseDataAttrs(elem)),
                         from: 0,
                         to: 1,
                     };